feat: clang-format: add format_all variant [DIP-243] (#107)
For our clang-format utility - we use git to obtain a list of files to
run the formatter against.

To speed up formatting and linting we attempt to get a subset of the
projects files. We currently use `git describe --tags`. This is a
reasonable heuristic - but it will not work properly if the project does
not have tags.

This PR adds a rule that will use `git ls-files` instead for projects
that do not have tags at the sacrifice of speed.