cc: toolchain: enable external_include_paths feature on mac
We require the external include paths feature to be able to
compile our code locally with -I to ensure compiler and linter
warnings are picked up in header files, and with -isystem when
the project is used externally to silence warnings when building
with the consuming projects options.

For some reason this feature is only enabled for the linux cc
toolchain configuration in the part of the toolchain config we
reuse from bazel (unix_cc_toolchain_config).

The only solution here is to vendor the unix_cc_toolchain_config
implementation and make the custom edits we require.