cc: quiet builds (#118)
Removes noise in our builds so we can focus on warnings and issues that
matter.

# Design Notes

## Remove default warnings from toolchain

The OTB bazel cc toolchain adds a few warnings to the compile line for
every file (i.e. `-Wall`).
We've kept this convention but it has made our builds extremely noisy,
particularly when building
third party code.

I've moved these options to only be added in our `swift_cc` wrappers.

## Silence implicit warnings when building third party code

Clang has a handful of warnings that are always on by default. I've
added a new feature to our
toolchain that disables these warnings when building third party code so
that builds aren't
spammed by warnings we're not likely to ever fix.

This has the downside of masking poor code hygiene in our dependencies,
but
the option can be overridden at the command line:
`bazel build --features=-no_default_warnings //...`

## Silence warnings about no symbols on macos

Libtool outputs a warning message if a file has no symbols. This shows
up quite
frequently in our builds - for instance in cases like `boringssl` where
entire files
are `#ifdef'd` away due to platform.

This is something we can ignore.

## Silence warnings about duplicate libraries on macos

This one is new with the latest linker upgrade. Now `ld` will issue a
warning if
for instance `-lpthread -lpthread` ends up on the link line.

This duplication is unavoidable in a big c++ codebase as you can't
assume someone
else is going to provide the dependency for you.

## Silence -Wno-deprecated-declarations in tests

This mostly comes from the extensive use of deprecated api's in gtest.
I've added a flag so that this can be re-enabled locally.

# Testing

I've tested the changes on both darwin platforms and linux and the
difference is
night and day.