feat(llvm-toolchain): support c++20 with llvm-14 and libc++ [IO-113] (#131)
Changes necessary to enable building C++17 and C++20 code with llvm-14.

# Background

The main challenge in implementing this is that the version of libc++
that ships with llvm-14 has several issues when it comes to building
c++17 and c++20 code:
- `memory_resource` implementation is still in the `experimental`
namespace
- There is a bug that caused a valid language feature to be deprecated
when using c++20 (`std::allocator<void>` specialization)
- libc++ is more aggressive in general about removing deprecated
language features, necessitating building with
`-D_LIBCPP_ENABLE_CXX20_REMOVED_FEATURES` if your code is still using
them.

# Design

The first major point is that `libstdc++` suffers from none of these
issues so I'm electing to make that the default for linux. On mac we
need to support libc++, and we would still like to build/test with this
configuration on linux, so most of these changes are in support of that.

I've chosen to use Bazel's
[features](https://bazel.build/docs/cc-toolchain-config-reference#features)
API for toolchain configuration to implement support for C++20.

Effectively what this enables us to do is all the user has to do is
specify there target is using `c++20`:
```bazel
cc_library(
    name = "cpp20",
    ...
    features = ["c++20"],
)
```

And underneath the hood, if we are using this specific toolchain, and
the user is building with `libc++` - then we will automatically add the
necessary definitions required to the compile line (only for the target
that requires it).

I've taken the opportunity to write a robust set of features for setting
language standards and implementation extensions so that we can move
toward using these `features` for our general language standard settings
(i.e. `swift_cc*` macros). This will enable us to remove the `c`
overloads of these functions in future releases.

I've also added a `relwdbg` feature as this is an oft requested
improvement to the build system.

These changes will cause a re-extraction/re-build of the codebase - so
it's best to pack as many of them in at once as possible.

This has also required a significant refactor of our toolchain internals
since previously we were not modelling the standard library choice as a
toolchain feature.

Lastly I've also added the patch to the llvm headers required to fix
using `std::allocator<void>`.