add mingw support [AP-2090] (#145)
Add MingW Cross Compiler Toolchain

Testing: The current consumer of this toolchain is Starling-Core.
This toolchain is limited to an x86_64 Linux host to build an x86_64
Windows target. With this toolchain, we are generating a binary and a
static library.

The artifacts were generated using CI/CD and were tested on Linux using
Wine and on a Windows 11 system.
The static library that was generated was also tested on Windows 11 to
generate the binary and static data was passed, and the results were
compared.

The same steps above were also done on an x86_64 Ubuntu 24.04 developer
machine.