Llvm: Remove `-fstack-protector` compiler flag [AP-3388]
In the LLVM toolchain configuration, the stack protector flag is
removed. This feature will emit a warning, if stack protection is not
deemed sensible for an array, which is the case for less than 8 bytes of
size. If warnings as errors is enabled, this will fail the build, if an
array of that size is defined.
We are removing the flag for our default builds, as stack protection
comes at a small runtime cost and we expect this default toolchain to be
used for development purposes only, where stack protection might not be
as relevant.
Users are encouraged to consider the stricter flags
`-fstack-protector-strong` or `-fstack-protector-all` for production builds.