Update protobuf to v25.1 (#152)
# Changes

I'm adding a new build flag which controls what C++ version it will use
across the entire codebase, including third party libraries. By default
Bazel uses C++14 (implicitly, even in third party libraries), which
normally won't be a problem if it wasn't for the fact that the newest
protobuf library which we are introducing didn't have have a problem
with this, specifically Abseil.

Repositories can now run:

```sh
$ bazel build --@rules_swiftnav//:global_cxx_standard=17 //...
```

Which compiles everything with the C++17 standard. Options for C++20 and
C++23 have been added as well.