Make toolchains bzl_mod compatible [IO-410] (#157)
In order to consume `rules_swiftnav` as `bzl_mod` with Bazel 7, the
toolchains need to be adapted.
- wrapper scripts now search also in folders, which belong to bazel mod.
- `sysroot`s are adapted and can be switched using a `config_setting`.
To enable the usage with `bzl_mod`, set
`--@rules_swiftnav//cc:enable_bzlmod=true`
- the executable `bin/llvm-as` does not exist. While this might work
with `WORKSPACE`, the build fails when using `bzl_mod`.

The backwards compatibility to `WORKSPACE` (Bazel 6) is kept and tested
as follows (see also linked PRs):
- [x] starling-core as `bzl_mod` (Bazel 7)
- [x] starling-core as `WORKSPACE`
- [x] starling as `WORKSPACE` 
- [x] orion-engine as `WORKSPACE`
- [x] orion as `WORKSPACE`