Migrate buildifier to MODULE [IO-410] (#163)
Migration of the tool `buildifier` from `WORKSPACE` to `MODULE.bazel`.
With this change the `rules_swiftnav` repo is migrated to Bazel 7,
however this does not affect any downstream repos, as tested in
`starling-core` with Bazel 6
([PR](https://github.com/swift-nav/starling-core/pull/1024)).
A test in `starling-core` with Bazel 7 is also
[successful](https://github.com/swift-nav/starling-core/pull/983).