Replace tilde with underscore in cc_static_library (#168)
When using `bzl_mod` the folder structure below
`bazel-out/k8-opt-aptiv/bin/external/` changes slightly. Modules
consumed as `bazel_dep` are located no longer in `foo` but in `foo~`.
The tilde (`~`) affects the archiver tooling of the `gcc-arm` toolchain,
when creating a `cc_static_library`, that uses a `.mri` script.

There is already an implementation, which replaces `+` with `_`. We can
do the same for the tilde.

Tests in
- [x] `starling` (Bazel 6/`WORKSPACE`)
- [x] `starling` (Bazel 7/`MODULE`)
- [x] `orion` (Bazel 6/`WORKSPACE`)