Allow usage of strip_include_prefix in cc_stamped_library (#179)
The use of includes in `cc_stamped_library` targets is incorrectly
propagating header paths with `-isystem`, which suppresses compiler
warnings in dependent targets. To resolve this, the `cc_stamped_library`
macro in `rules_swiftnav` must be updated to support
`strip_include_prefix` and remove its dependency on the `includes`
attribute.

Tested in 
- [x] starling-core
- [x] starling